fix(clippy): restore the Full gate after RV16 extractions
Remove 20 orphaned helpers left by the dataset modularization, gate 42
test-only helpers behind #[cfg(test)], and apply clippy's redundant
closure/borrow/cast fixes. Tauri routes, JSON contracts and behavior
are unchanged.

+++ src-tauri/src/remote_databases.rs
@@ -286,7 +286,7 @@ where
                 }
             }
         } else {
-            null_count = column.null_count() as usize;
+            null_count = column.null_count();
         }
         ensure_not_cancelled(is_cancelled)?;
         input_columns.push(RemoteInputColumn {
